cables/requirements.txt
aztec-code-generator==0.11
Pillow==11.0.0
qrcode==7.4.2
